Align E2B provider module boundaries
