+++ src/setup/e2b/index.ts
@@ -1,0 +1,8 @@
+export type { BuildSelfBenchE2BTemplateOptions, E2BTemplateBuildApi } from "./build.js";
+export { buildSelfBenchE2BTemplate } from "./build.js";
+export { setupE2B } from "./cli.js";
+export {
+  normalizeE2BBuildName,
+  normalizeE2BDomain,
+  normalizeE2BTemplateReference,
+} from "./template.js";
