+++ src/setup/e2b/template.ts
